refactor: extract private PDF fidelity evidence helpers
